Rename viewdsl-design to viewdsl-material
This breaks the public API since the package is also renamed, but it's
not a big deal compared to AndroidX migration, and we're still in alpha.
