+++ sample/src/main/java/com/louiscad/splittiessample/about/AboutUi.kt
@@ -32,9 +32,9 @@ import splitties.viewdsl.core.Ui
 import splitties.viewdsl.core.add
 import splitties.viewdsl.core.v
 import splitties.viewdsl.core.wrapContent
-import splitties.viewdsl.design.appBarLParams
-import splitties.viewdsl.design.contentScrollingWithAppBarLParams
-import splitties.viewdsl.design.defaultLParams
+import splitties.viewdsl.material.appBarLParams
+import splitties.viewdsl.material.contentScrollingWithAppBarLParams
+import splitties.viewdsl.material.defaultLParams
 import splitties.views.appcompat.Toolbar
 import splitties.views.centerText
 import splitties.views.textAppearance
