+++ sample/src/main/java/com/louiscad/splittiessample/main/MainUi.kt
@@ -38,13 +38,13 @@ import splitties.viewdsl.core.margin
 import splitties.viewdsl.core.matchParent
 import splitties.viewdsl.core.v
 import splitties.viewdsl.core.verticalLayout
-import splitties.viewdsl.design.EXIT_UNTIL_COLLAPSED
-import splitties.viewdsl.design.PIN
-import splitties.viewdsl.design.SCROLL
-import splitties.viewdsl.design.actionBarLParams
-import splitties.viewdsl.design.appBarLParams
-import splitties.viewdsl.design.contentScrollingWithAppBarLParams
-import splitties.viewdsl.design.defaultLParams
+import splitties.viewdsl.material.EXIT_UNTIL_COLLAPSED
+import splitties.viewdsl.material.PIN
+import splitties.viewdsl.material.SCROLL
+import splitties.viewdsl.material.actionBarLParams
+import splitties.viewdsl.material.appBarLParams
+import splitties.viewdsl.material.contentScrollingWithAppBarLParams
+import splitties.viewdsl.material.defaultLParams
 import splitties.viewdsl.recyclerview.experimental.setSingleView
 import splitties.viewdsl.recyclerview.recyclerView
 import splitties.views.appcompat.Toolbar
