+++ viewdsl-material/src/main/AndroidManifest.xml
@@ -14,4 +14,4 @@
   ~ limitations under the License.
   -->
 
-<manifest package="splitties.viewdsl.design"/>
+<manifest package="splitties.viewdsl.material"/>
