+++ viewdsl-material/src/main/java/splitties/viewdsl/material/TextInputLayout.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package splitties.viewdsl.design
+package splitties.viewdsl.material
 
 import androidx.annotation.IdRes
 import com.google.android.material.textfield.TextInputEditText
